refactor: eventsReducer, ScheduleTable, Main components
