+++ src/components/ScheduleTable/EventHideButton/EventHideButton.js
