+++ src/components/ScheduleTable/EventHideButton/index.js
