+++ src/constants/actions-types.js
@@ -17,4 +17,5 @@ export const REMOVE_EVENT = 'EVENTS/REMOVE_EVENT';
 export const SET_EVENT_TYPE_FILTER = 'EVENTS/SET_EVENT_TYPE_FILTER';
 export const LOADED_ORGANIZERS = 'APP/LOADED_ORGANIZERS';
 export const SET_HIDDEN_EVENTS = 'EVENTS/SET_HIDDEN_EVENTS';
+export const REMOVE_HIDDEN_EVENT = 'EVENTS/REMOVE_HIDDEN_EVENT';
 export const SWITCH_VISIBILITY_HIDDEN_EVENTS = 'APP/SWITCH_VISIBILITY_HIDDEN_EVENTS';
