+++ src/reducers/eventsReducer.js
@@ -1,6 +1,7 @@
 import {
   LOADED_SCHEDULE,
   REMOVE_EVENT,
+  REMOVE_HIDDEN_EVENT,
   SET_EVENT_TYPE_FILTER,
   SET_HIDDEN_EVENTS,
 } from '../constants/actions-types';
@@ -10,6 +11,7 @@ const initialState = {
   hiddenEvents: JSON.parse(localStorage.getItem('hiddenEvents')) || [],
   eventTypeFilter: JSON.parse(localStorage.getItem('eventTypeFilter')) || [],
 };
+let newHiddenEvents;
 
 const eventsReducer = (state = initialState, {type, payload}) => {
   switch (type) {
@@ -20,7 +22,13 @@ const eventsReducer = (state = initialState, {type, payload}) => {
     case SET_EVENT_TYPE_FILTER:
       return {...state, eventTypeFilter: payload};
     case SET_HIDDEN_EVENTS:
-      return {...state, hiddenEvents: payload};
+      newHiddenEvents = Array.from(new Set([...state.hiddenEvents, ...payload]));
+      localStorage.setItem('hiddenEvents', JSON.stringify(newHiddenEvents));
+      return {...state, hiddenEvents: newHiddenEvents};
+    case REMOVE_HIDDEN_EVENT:
+      newHiddenEvents = state.hiddenEvents.filter(evt => evt !== payload);
+      localStorage.setItem('hiddenEvents', JSON.stringify(newHiddenEvents));
+      return {...state, hiddenEvents: newHiddenEvents};
     default:
       return state;
   }
